combine all into the app.jsx file, changed navbar, cleaned up buttons for crud and the modals
