+++ public/less/main.less
@@ -1,6 +1,6 @@
+@import "./profile.less";
 @primary: #00F1AE; //green
 @secondary: #F6ECDA; //white-gray
-@import "./profile.less";
 @blue: #3399ff;
 @red: #ff0033;
 @purple: #5c2751;
@@ -15,6 +15,10 @@ html {
 }
 h3 {
   margin-top: 0;
+}
+
+.btn-action {
+  background: @blue;
 }
 
 .logo {
@@ -58,9 +62,22 @@ h3 {
   width: 150px;
 }
 
+#navbar {
+   display: -webkit-flex;
+   display: flex;
+   width: 100%;
+   ul {
+    width: 300px;
+   }
+   .spacer{
+    -webkit-flex: 1;
+            flex: 1;
+   }
+}
+
  body {
-   min-height: 2000px;
-   padding-top: 70px;
+   margin-top: 55px !important;
+   // padding-top: 70px;
  }
  .head{
    display: -webkit-flex;
@@ -178,7 +195,7 @@ h3 {
   border-radius: @radius;
 }
 
-.dashHeader{
+.profileHeader{
   display: -webkit-flex;
   display: flex;
   // display: inline-block;
@@ -189,10 +206,25 @@ h3 {
             flex: 1;
   }
   .editRound {
-    border-radius: @radius;
+    width: 40px;
+    border-radius: 100%;
+    height: 40px;
+    padding: 0;
+    background: @blue;
   }
 }
 
+.eventsHeader{
+  display: -webkit-flex;
+  display: flex;
+  // display: inline-block;
+  margin: 0;
+  width: 100%;
+  .spacer {
+    -webkit-flex: 1;
+            flex: 1;
+  }
+}
 
 .fullModal{
   .imgPreview{
@@ -255,6 +287,7 @@ h3 {
         min-height: 7rem;
       }
       button{
+        background: @blue;
         margin-top: 2rem;
         margin-left: 23rem;
         border-radius: @radius;
